+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 # Changelog
 
-All notable changes to atex will be documented in this file.
+All notable changes to elixir-dasl will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased] -->
+## [Unreleased]
+
+### Added
+
+- `~CID` sigil for easily constructing known CIDs.
+- New inspect string for CIDs using the new sigil.
 
 ## [0.1.0] - 2026-04-08